+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - fix bug in `MovementByJoystick` when called `idle` method.
 - `positionsItselfAndKeepDistance` little improvements.
 - `AutomaticRandomMovement` improvements. Adds param `direction` to decide which direction could be the movement.
+- Adds `moveToPosition` in `Movement` mixin.
 
 # 3.0.0-beta.11